DPR2-69 integration of the library working in the reporting-mi project
