--- src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportingmi/data/ProductDefinitionRepository.kt
@@ -1,8 +1,0 @@
-package uk.gov.justice.digital.hmpps.digitalprisonreportingmi.data
-
-import uk.gov.justice.digital.hmpps.digitalprisonreportingmi.data.model.ProductDefinition
-
-interface ProductDefinitionRepository {
-
-  fun getProductDefinitions(): List<ProductDefinition>
-}
