--- src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportingmi/exception/DuplicateDataSetIdException.kt
@@ -1,3 +1,0 @@
-package uk.gov.justice.digital.hmpps.digitalprisonreportingmi.exception
-
-class DuplicateDataSetIdException(val dataSetId: String) : RuntimeException("Error: Duplicate dataSet Id found: $dataSetId")
