+++ src/test/resources/application-test.yml
@@ -15,7 +15,7 @@ spring:
     oauth2:
       resourceserver:
         jwt:
-          jwk-set-uri: ${hmpps.auth.url}/.well-known/jwks.json
+          jwk-set-uri: https://abc/.well-known/jwks.json
     user:
       roles: ROLE_PRISONS_REPORTING_USER
